Reference

Freenect.DepthFormat

Freenect.LEDMode

Freenect.RawTiltState

Freenect.Resolution

Freenect.TiltStatusCode

Freenect.VideoFormat

Freenect.sync_get_depth

Freenect.sync_get_depth

Freenect.sync_get_depth_with_res

Freenect.sync_get_depth_with_res

Freenect.sync_get_pointcloud

Freenect.sync_get_tilt_state

Freenect.sync_get_tilt_state

Freenect.sync_get_video

Freenect.sync_get_video

Freenect.sync_get_video_with_res

Freenect.sync_get_video_with_res

Freenect.sync_set_led

Freenect.sync_set_tilt_degs

Freenect.DepthFormat — Type

See http://openkinect.org/wiki/Protocol_Documentation#RGB_Camera for more information.

Freenect.LEDMode — Type

Available modes for the LED on the Kinect.

Freenect.RawTiltState — Type

struct RawTiltState

accelerometer_x::Int16

accelerometer_y::Int16

accelerometer_z::Int16

tilt_angle::Int8

tilt_status::Int8

This data is currently uninterpreted and only provided raw.

Freenect.Resolution — Type

Enumeration of available resolutions. Not all available resolutions are actually supported for all video formats. Frame modes may not perfectly match resolutions. For instance, resolution_medium is 640x488 for the IR camera.

Freenect.TiltStatusCode — Type

Possible status codes returned in RawTiltState.

Freenect.VideoFormat — Type

See http://openkinect.org/wiki/Protocol_Documentation#RGB_Camera for more information.

Freenect.sync_get_depth — Method

sync_get_depth(depth::Any, timestamp::Any, index::Any, fmt::Any) -> Int32

Does the exact same as sync_get_depth_with_res, with a default resolution of resolution_medium.

Freenect.sync_get_depth — Method

sync_get_depth(index::Integer, fmt::DepthFormat) -> Tuple{Array{UInt16,2},Int64}

Synchronous depth function with resolution resolution_medium, starts the runloop if it isn't running.

The returned array is copied before returning and is safe to store.

Freenect.sync_get_depth_with_res — Method

sync_get_depth_with_res(depth::Any, timestamp::Any, index::Any, res::Any, fmt::Any) -> Int32

Synchronous depth function, starts the runloop if it isn't running

In the raw pointer version, the returned buffer is valid until this function is called again, after which the buffer must not be used again. Make a copy if the data is required. The version returning an array does not have this limitation.

Freenect.sync_get_depth_with_res — Method

sync_get_depth_with_res(index::Integer, res::Resolution, fmt::DepthFormat) -> Tuple{Array{UInt16,2},Int64}

Synchronous depth function, starts the runloop if it isn't running.

The returned array is copied before returning and is safe to store.

Freenect.sync_get_pointcloud — Method

sync_get_pointcloud(index::Integer) -> Tuple{Array{Float64,3},Int64}

Helper for getting a pointcloud from the camera at index.

The resulting point cloud is relative to the camera, with X forward, Y left, and Z up.

This function uses a fixed homography matrix - you may have better results for your own hardware by calibrating your Kinect.

Freenect.sync_get_tilt_state — Method

sync_get_tilt_state(state::Any, index::Any) -> Int32

Tilt state function, starts the runloop if it isn't running.

The returned pointer is only safe until the next call to this function.

Freenect.sync_get_tilt_state — Method

sync_get_tilt_state(index::Integer) -> RawTiltState

Tilt state function, starts the runloop if it isn't running.

The returned RawTiltState struct is safe to store.

Freenect.sync_get_video — Method

sync_get_video(video::Any, timestamp::Any, index::Any, fmt::Any) -> Int32

Does the exact same as sync_get_video_with_res, with a default resolution of resolution_medium.

Freenect.sync_get_video — Method

sync_get_video(index::Integer, fmt::VideoFormat) -> Tuple{Any,Int64}

Synchronous video function with resolution resolution_medium, starts the runloop if it isn't running.

The returned array is copied before returning and is safe to store.

Freenect.sync_get_video_with_res — Method

sync_get_video_with_res(video::Any, timestamp::Any, index::Any, res::Any, fmt::Any) -> Int32

Synchronous video function, starts the runloop if it isn't running

The returned buffer is valid until this function is called again, after which the buffer must not be used again. Make a copy if the data is required.

Freenect.sync_get_video_with_res — Method

sync_get_video_with_res(index::Integer, res::Resolution, fmt::VideoFormat) -> Tuple{Any,Int64}

Synchronous video function, starts the runloop if it isn't running.

The returned array is copied before returning and is safe to store.

Freenect.sync_set_led — Method

sync_set_led(led::Any, index::Any) -> Int32

Set the LED to the given mode, starts the runloop if it isn't running.

Freenect.sync_set_tilt_degs — Method

sync_set_tilt_degs(angle::Any, index::Any) -> Int32

Tilt the kinect to angle. Starts the runloop if it isn't running.